Avoid Xcode warnings.

+++ BikeAirQualitySensorsiOS/BLE/BLE.m
@@ -305,8 +305,8 @@ static int rssi = 0;
 {
     char b1[16];
     char b2[16];
-    [UUID1.data getBytes:b1 length:b1];
-    [UUID2.data getBytes:b2 length:b2];
+    [UUID1.data getBytes:b1 length:(unsigned int)b1];
+    [UUID2.data getBytes:b2 length:(unsigned int)b2];
     
     if (memcmp(b1, b2, UUID1.data.length) == 0)
         return 1;
@@ -318,7 +318,7 @@ static int rssi = 0;
 {
     char b1[16];
     
-    [UUID1.data getBytes:b1 length:b1];
+    [UUID1.data getBytes:b1 length:(unsigned int)b1];
     UInt16 b2 = [self swap:UUID2];
     
     if (memcmp(b1, (char *)&b2, 2) == 0)
@@ -330,7 +330,7 @@ static int rssi = 0;
 -(UInt16) CBUUIDToInt:(CBUUID *) UUID
 {
     char b1[16];
-    [UUID.data getBytes:b1 length:b1];
+    [UUID.data getBytes:b1 length:(unsigned int)b1];
     return ((b1[0] << 8) | b1[1]);
 }
